api/requirements.txt
appdirs==1.4.3
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
cli-helpers==1.2.1
click==7.1.1
configobj==5.0.6
distlib==0.3.0
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
filelock==3.0.12
humanize==2.3.0
idna==2.9
pgcli==3.0.0
pgspecial==1.11.9
Pillow==7.1.1
postgres==3.0.0
prompt-toolkit==3.0.5
psycopg2==2.8.5
psycopg2-binary==2.8.5
psycopg2-pool==1.1
Pygments==2.6.1
python-decouple==3.3
pytz==2019.3
requests==2.23.0
setproctitle==1.1.10
six==1.14.0
sqlparse==0.3.1
tabulate==0.8.7
terminaltables==3.1.0
urllib3==1.25.8
virtualenv==20.0.17
wcwidth==0.1.9
webencodings==0.5.1
